feat: createRoom 로직 분리 (#128)
